feature: Add a distinct node sort
Start of the code to allow distinct node sorting, this currently is
using the nodes pk as the order, but need extending to be any of the
applied dag attributes. ie  dag_sequence_path/ dag_node_path etc.

This will enable a  'Depth first, preorder' sort where each node is only
included the first time it is visited.